requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
black==24.4.2
boto3==1.34.141
botocore==1.34.141
CacheControl==0.14.0
cachetools==5.4.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.8
daphne==4.1.2
Django==5.0.7
django-celery-beat==2.6.0
pytz
django-celery-results==2.5.1
django-channels==0.7.0
django-cors-headers==4.4.0
django-storages==1.14.4
django-timezone-field==6.1.0
djangorestframework==3.15.2
firebase-admin==6.5.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
gunicorn==22.0.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
jmespath==1.0.1
kombu==5.3.7
msgpack==1.0.8
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
Pillow==9.0.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.7
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
stripe==10.2.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
zope.interface==6.4.post2
ecdsa